Small improvement to Mist's day/night handling
Saves an atomicboolean inside the map instead of a normal one, now you can just use atomicboolean.set(false) instead of replacing the whole entry. I also fixed a potential bug that would let waterbenders have a night boost in the nether or in the end, this occurs because !isDay is used instead of isNight, for firebenders this isn't a problem as it's always day in the nether and will still give them a boost, but with waterbenders it'll give them a boost while it shouldn't. Same goes for the end world. I still think saving a world object is unnecessary as you can use a string too, as there cant be 2 worlds with the same name.